マップチップ最低限

+++ ApplicationCode/MapChip/ChipData.h
@@ -27,8 +27,8 @@ public:
 	XMFLOAT2 GetPos() { return pos_; }
 	int GetChipNumber() { return chipNumber; }
 
-	int GetSize() { return size_.size(); }
-	int GetHalfSize() { return size_.size()/2; }
+	int GetSize() { return (int)size_.size(); }
+	int GetHalfSize() { return (int)size_.size()/2; }
 
 	bool GetDrawFlag() { return drawFlag; }
 	void SetDrawFlag(bool flag) { this->drawFlag = flag; }
